Fix header for directive template
Change the header position in the directive template.
When the header is at the top level it counts as a node root in the DOM and throw an error in angularjs.
src: https://docs.angularjs.org/error/$compile/tplrt
This issue as been fixed in Angularjs 1.4 beta, so we won't have problems in the future (currently 1.3.8)

Author: Damien Corneau <corneadoug@gmail.com>

Closes #29 from corneadoug/fix/licenseInDirectivesViews and squashes the following commits:

deb5a61 [Damien Corneau] Fix header for directive template