getPort should not have curly bracket in its body
This patch simply remove the usage of curly bracket in the body of the getPort method of the app.js.
This is needed so the AppScriptServlet can do its correctly in case of developement in both client and server sides.

Author: Eric Charles <eric@datalayer.io>

Closes #127 from echarles/app-js-server-mode and squashes the following commits:

dfaba55 [Eric Charles] getPort should not have curly bracket in its body