feat: use ZEPPELIN_HOSTNAME to specify hostname
ZEPPELIN_HOSTNAME supports running Zeppelin in environments where the
hostname of the server does not match the hostname delivered to the
browser. If ZEPPELIN_HOSTNAME is not set, java.net.InetAddress.
getLocalHost().getHostName() is still used to lookup the server's
current hostname.

NotebookServerTests was renamed to NotebookServerTest to permit the
test cases to run using Maven's test goal.