Fix time-dependant scheduling test
test-schedule can fail when it starts late in a second, as the result will then return the next second due to in-system delays (particularly cron not reacting in time to config.put(cron, null)).
Adding an additional waiting second before measuring dateFinished allows us to take a value that is independent of a possible lagging execution, waiting another second makes sure that the process has properly stopped.

Alternatively, we can test the time and make sure to not write to cron when just before the full second - that would be quicker (less waiting) but slightly more hacky.

Please review.