ZEPPELIN-319 optionally wait an additional .5 seconds, if assertion can be expected to fail
This test failed randomly during a build. To make sure this is not due to timing issues, we can optionally wait an additional .5 seconds.
This is a band-aid around bad design, but I'm pragmatically trying to get the build to be more stable.

Please review, and if you can come up with a better designed test we should discuss it here or in the issue.