[ZEPPELIN-319]wait an additional .5 seconds, if assertion will fail
This test failed randomly during a build. To make sure this is not due to timing issues, we can optionally wait an additional .5 seconds.
This is a band-aid around bad design, but I'm pragmatically trying to get the build to be more stable. This should also help determine if there is an actual (randomly striking) issue in the underlying procedure being tested, by making the test more reliable.

Please review, and if you can come up with a better designed test (and avoid waiting for threads to finish in a fixed time interval) we should discuss it here or in the issue.

Author: Rick Moritz <rahvin@gmail.com>

Closes #323 from RPCMoritz/ZEPPELIN-319 and squashes the following commits:

27b6608 [Rick Moritz] Switched to isTerminated() for clearer semantics
5e4b8d3 [Rick Moritz] added assertEquals import
10fd22d [Rick Moritz] Fix build and cosmetics
0455a66 [Rick Moritz] Refactoring and moving to job.isRunning check
528dbfa [Rick Moritz] ZEPPELIN-319 optionally wait an additional .5 seconds, if assertion can be expected to fail