Restore Navbar connection status for websocket only
### What is this PR for?
After the Shiro PR was merged, the Navbar status was changed to include the token status/name, however that part is reserved to show the Websocket status. Because of that it usually show a green circle with Disconnected next to it.
This PR sets the status to Websocket only.

### What type of PR is it?
Bug Fix

### Is there a relevant Jira issue?
No

### How should this be tested?
You shouldn't see see Disconnected next to a green round in the navbar anymore

### Screenshots (if appropriate)
Bug:
<img width="386" alt="screen shot 2016-01-13 at 12 02 24 pm" src="https://cloud.githubusercontent.com/assets/6766992/12283801/927d05e8-b9ed-11e5-8911-9725b4a7c94a.png">

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: Sejun Ra <sejun@nflabs.com>

Closes #628 from sejunra/fix/NavbarConnectedStatus and squashes the following commits:

ee8a2fd [Sejun Ra] Restore Navbar connection status for websocket only