ZEPPELiN-654 Improvement to SparkSqlInterpreter
### What is this PR for?
Improve performance of Sparksqlinterpreter
Use StringBuilder instead of String when building the results returned by SparksqlInterpreter

### What type of PR is it?
Improvement

### Todos
NA

### Is there a relevant Jira issue?

### How should this be tested?
Try to run a Sparksql command that returns large number of rows(make sure `zeppelin.spark.maxResult` is set to a larger value) with and without fix.
Keep note of the time taken in each case.
The time taken with this change  is significantly lesser.

### Screenshots (if appropriate)
NA

### Questions:
* Does the licenses files need update?  No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: karuppayya <karuppayya1990@gmail.com>

Closes #695 from karup1990/sqlc-imp and squashes the following commits:

11789e8 [karuppayya] use stringbuilder instead of string