Simple improvement to showDF() to use StringBuilder
What is this PR for?

The way tables of results are currently getting built within
the ZeppelinContext is pretty slow due to way strings are
getting constructed.

What type of PR is it?

Performance improvement.

How should this be tested?

The current tests should catch any issues introduced by it.