[ZEPPELIN-658] Scala: not accepting companion objects if defined in different lines
### What is this PR for?
In Scala, a singleton companion object can be defined along the class. Because of the way code lines are split, the interpreter is not passing the companion object along with the class code, or indicating any possible line continuation (eg. leaving out '}'). This causes an error with the Scala REPL/compiler.

### What type of PR is it?
Bug Fix

### Todos
* [x] - Fix line splitting
* [x] - Add tests

### What is the Jira issue?
https://issues.apache.org/jira/browse/ZEPPELIN-658
https://issues.apache.org/jira/browse/ZEPPELIN-747

### How should this be tested?
Add a new paragraph with Scala/Spark interpreter and define a class with its companion object.
See bug for more details

### Screenshots (if appropriate)
N/A

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: Felix Cheung <felixcheung_m@hotmail.com>

Closes #780 from felixcheung/scalarepl and squashes the following commits:

5087dae [Felix Cheung] check for continuation