increase implicitlyWait for WebDriver
### What is this PR for?
This is with respect to travis build failure on `sendKeys`.

An implicit wait is to tell WebDriver to poll the DOM for a certain amount of time when trying to find an element or elements if they are not immediately available. The default setting is 0. Once set, the implicit wait is set for the life of the WebDriver object instance.

#### Update
Travis mostly fails on `sendKeys`; hence, change the way text is sent to paragraph.

### What type of PR is it?
Hot Fix

### What is the Jira issue?
N/A

### How should this be tested?
CI should be green.

### Screenshots (if appropriate)

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: Prabhjyot Singh <prabhjyotsingh@gmail.com>

Closes #785 from prabhjyotsingh/increaseImplicitlyWait and squashes the following commits:

5c36fac [Prabhjyot Singh] add waitForParagraph(1, "FINISHED"); change assertTrue to waitForParagraph(1, "FINISHED");
0249b49 [Prabhjyot Singh] replace sendkeys with javascript function
734a901 [Prabhjyot Singh] increase implicitlyWait for WebDriver