[ZEPPELIN-1983] Feedback about success/failure when interpreter is re…
### What is this PR for?
When a interpreter is restarted, there is no feedback to user whether the action is received/honoured by zeppelin. This might also lead them to click restart multiple times, before they realize it has taken affect.
Also fixed a null bug.

### What type of PR is it?
[Improvement + Bug fix]

### Todos
* None

### What is the Jira issue?
ZEPPELIN-1983

### How should this be tested?
Restart a interpreter to test the positive case. The message should disappers on its own after sometime if not closed. A simple way to test the negative case is to stop the server and then try restarting.

### Screenshots (if appropriate)
#### Positive case
![positive case](https://cloud.githubusercontent.com/assets/4542030/22111327/fa675304-de84-11e6-9323-da49f2f70902.png)

#### Negative case
![negative case](https://cloud.githubusercontent.com/assets/4542030/22111326/fa31bd84-de84-11e6-8a2e-07aaf0179275.png)

### Questions:
* Does the licenses files need update?
-No
* Is there breaking changes for older versions?
-No
* Does this needs documentation?
-No

Author: Beria <beria@qubole.com>

Closes #1914 from beriaanirudh/ZEPPELIN-1983 and squashes the following commits:

761b986 [Beria] [ZEPPELIN-1983] Feedback about success/failure when interpreter is restarted