ZEPPELIN-844: Navbar menu should not show up if user is not authenticated
### What is this PR for?
If a user in not authenticated, s/he should not see Navbar menu, or search bar, or import/create note book links.

### What type of PR is it?
[Bug Fix]

### Todos
* [x] - Hide all option from frontend

### What is the Jira issue?
[ZEPPELIN-844](https://issues.apache.org/jira/browse/ZEPPELIN-844)

### How should this be tested?
Change shiro conf to enable authentication. Now when home page is loaded you should not see any of the Navbar menu or option to create or import notebook

### Screenshots (if appropriate)
Before:
<img width="1439" alt="screen shot 2016-05-10 at 3 08 10 pm" src="https://cloud.githubusercontent.com/assets/674497/15142088/1918535c-16c1-11e6-9c2e-2662fdfb90e6.png">

After:
<img width="1440" alt="screen shot 2016-05-10 at 2 49 05 pm" src="https://cloud.githubusercontent.com/assets/674497/15142087/19183778-16c1-11e6-896e-1dc708c8700b.png">

### Questions:
* Does the licenses files need update? no
* Is there breaking changes for older versions? no
* Does this needs documentation? no

Author: Prabhjyot Singh <prabhjyotsingh@gmail.com>

Closes #881 from prabhjyotsingh/ZEPPELIN-844 and squashes the following commits:

c7fdeaf [Prabhjyot Singh] check the content of ticket, and based on it show or hide logout button.
340f7ad [Prabhjyot Singh] tune logout functionality
d865248 [Prabhjyot Singh] ZEPPELIN-844: Navbar menu should not show up if user is not authenticated.