[ZEPPELIN-846] Fix Long Notebook Dropdown
### What is this PR for?
This PR is fixing the bug by implementing a new layout for the Notebook Dropdown menu.
Reason being that setting a maximum height and having a scrollbar doesn't mix with having bootstrap submenus.

In term of dropdown layout, I think we have 2 options:
* The one I implemented, a tree view similar to the one in the home page, I can probably uniform both of them later. One possible problem is the dropdown width getting bigger and bigger as we indent at each submenu.
* The second one would not have any indent, but use a background color to show elements in the same folder (A grey that goes clearer as we go in the submenus)

<img width="691" alt="screen shot 2016-06-25 at 1 42 41 am" src="https://cloud.githubusercontent.com/assets/710411/16344155/2d387af0-3a76-11e6-87ba-9626d9f17fea.png">

Reworking the whole notebook list and dropdown will probably need a lot more work later (especially in term of refactoring and code cleaning), so I'm not sure where the scope of this PR should stop, especially with the release coming. I think the layout I implemented is fine to solve the immediate problem, and it is always possible to switch to the layout of option 2 later if we think it is best.

### What type of PR is it?
Hot Fix

### Todos
* [ ] - Rebase & Double Check the filtering and Notebook ordering after #1066

### What is the Jira issue?
https://issues.apache.org/jira/browse/ZEPPELIN-846

### How should this be tested?
Just play with the folder structure and check the notebook dropdown

### Screenshots (if appropriate)
![openingfolder](https://cloud.githubusercontent.com/assets/710411/16343658/583e1064-3a73-11e6-8607-bcdced587636.gif)

![notebookisrunnable](https://cloud.githubusercontent.com/assets/710411/16343663/5c10894c-3a73-11e6-9f54-c32136d3437a.gif)

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: Damien CORNEAU <corneadoug@gmail.com>

Closes #1081 from corneadoug/fix/longNavbarNotebookList and squashes the following commits:

223a7d3 [Damien CORNEAU] Fix Notebook dropdown css when on mobile
8170782 [Damien CORNEAU] Fix Notebook ordering in dropdown
5a0fa09 [Damien CORNEAU] fix jshint errors
a52b219 [Damien CORNEAU] Change Notebook dropdown style

(cherry picked from commit eca39e2f3f3082bf997183879e5999b4f61faf8b)
Signed-off-by: Mina Lee <minalee@apache.org>