Block all routes except Home when user isn't logged
### What is this PR for?
When user isn't logged while the auth is activated, we are hiding the notebooks and most of the menus.
However you can still access them using the url routes.
This PR is blocking the url routes except the home route ('/') in that case.

### What type of PR is it?
Improvement

### What is the Jira issue?
https://issues.apache.org/jira/browse/ZEPPELIN-1123

### How should this be tested?
* Activate the auth in shiro.ini
* Launch Zeppelin, and try to visit `localhost:8080/#/interpreter` by typing it in your browser
* You should be redirected to '/'

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: Damien CORNEAU <corneadoug@gmail.com>

Closes #1153 from corneadoug/ZEPPELIN-1123 and squashes the following commits:

58c5a6e [Damien CORNEAU] Remove preventDefault
ebb9136 [Damien CORNEAU] Block all routes except Home when user isn't logged