[ZEPPELIN-1316] Zeppelin can not start due to an incorrect Interpreter Setting
### What is this PR for?
If there are problems with the information defined in the Interpreter Setting (conf / interpreter.json), Interpreter Group during the defined file does not exist, can not run Zeppelin.
Thus, it was an exception handling for that part.

In order to address this issue in the current version it was restarted after deleting Interpreter.json.

### What type of PR is it?
Bug Fix

### Todos
- [x] - check to Interpreter Setting is null

### What is the Jira issue?
https://issues.apache.org/jira/browse/ZEPPELIN-1316

### How should this be tested?

1. Modify the following part of the Interpreter.json file.
```json
{
      "interpreterGroup": [
        {
          "name": "spark",
          "class": "fake and invalid class path more more...", // invalid value
          "defaultInterpreter": false
        }
}
```

2.
After restarting the Zeppelin, Zeppelin Confirm that normally works.
Version of the PR is to be normally started up.
It should also be display all possible Interpreter Web UI.

### Screenshots (if appropriate)
#### before
<img width="618" alt="error" src="https://cloud.githubusercontent.com/assets/10525473/17543783/104363e4-5f0e-11e6-8279-e018a708ccb8.png">

#### after
<img width="1026" alt="correct" src="https://cloud.githubusercontent.com/assets/10525473/17543867/899da484-5f0e-11e6-9f8b-b1bac7cb2718.png">

### Questions:
* Does the licenses files need update? no
* Is there breaking changes for older versions? no
* Does this needs documentation? no

Author: CloverHearts <cloverheartsdev@gmail.com>
Author: CloverHearts <cloverhearts@CloverHeartsui-MacBook-Pro.local>

Closes #1311 from cloverhearts/ZEPPELIN-1316 and squashes the following commits:

c960be2 [CloverHearts] Merge branch 'master' into ZEPPELIN-1316
81eb13a [CloverHearts] Merge branch 'master' into ZEPPELIN-1316
679afbb [CloverHearts] chante to information log to warning log in interpreterfactory
7aa7643 [CloverHearts] Modification - log typo
d22358b [CloverHearts] check to Interpreter Setting is null on interpreter factory in loadfromfile method