[ZEPPELIN-1359] Commit correctly ordered karma.conf file
### What is this PR for?
In the karma-conf.js file, dependencies are usually alphabetically ordered.
On a clean repo build, the file is changed to be alphabetically order those dependencies.
SInce one of those is not in the right order, it creates a modified file in the repository after build.

This PR is simply re-ordering the file so that there is no modifications on build.

### What type of PR is it?
Bug Fix

### What is the Jira issue?
https://issues.apache.org/jira/browse/ZEPPELIN-1359

### How should this be tested?
Build the source code from a clean repo, there shouldn't be any file changed if you type `git status`

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: Damien CORNEAU <corneadoug@gmail.com>

Closes #1355 from corneadoug/ZEPPELIN-1359 and squashes the following commits:

6d9a8da [Damien CORNEAU] Add correct karma.conf.js