Update Zeppelin Website building instructions
### What is this PR for?
Usually in order to push jekyll code to github pages, it will compile using the `--safe`
One drawback is that you cannot use any plugins.

After working on PR #1356 and introduction a plugin, using the build line with `--safe` specified in the README is failing.

However in our case, we are not forced to use that flag, since we are sending the compiled website directly to the Apache SVN, and I was able to confirm that fact when I sent the updated website to the SVN and it was rendered correctly

Also since plugins like Google Analytics needs to be activated only in production we are adding a special production environment option to add `JEKYLL_ENV=production`

### What type of PR is it?
Documentation

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: Damien CORNEAU <corneadoug@gmail.com>

Closes #1405 from corneadoug/update/buildCommandLine and squashes the following commits:

a5652b1 [Damien CORNEAU] Change also in the website contribution guide
08b46e5 [Damien CORNEAU] Update to handle the production flag
acd0b04 [Damien CORNEAU] Remove --safe from building instruction