ZEPPELIN-1473. It is not necessary to create SQLContext in LivyInterpreter
### What is this PR for?
Livy will create SQLContext/HiveContext internally, (LIVY-94), so it is not necessary to create that in LivyInterpreter. Otherwise sqlContext in zeppelin will override that in livy.

### What type of PR is it?
[Bug Fix]

### Todos
* [ ] - Task

### What is the Jira issue?
* https://issues.apache.org/jira/browse/ZEPPELIN-1473

### How should this be tested?
Tested manually.  HiveContext is created properly in livy (with proper livy configuration), and can access hive data.

### Screenshots (if appropriate)
![image](https://cloud.githubusercontent.com/assets/164491/18743886/bff7ae8e-80ed-11e6-83e6-0769c30e4094.png)

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: Jeff Zhang <zjffdu@apache.org>

Closes #1450 from zjffdu/ZEPPELIN-1473 and squashes the following commits:

bae46a1 [Jeff Zhang] ZEPPELIN-1473. It is not necessary to create SQLContext in LivyInterpreter