[ZEPPELIN-1632] Add the possibility to cancel flink jobs in local mode
### What is this PR for?
Especially for long running flink streaming jobs it should be possible to cancel a job by clicking the cancel button in the zeppelin ui. This pull request contains an implementation to cancel jobs when the flink interpreter is running in local mode.

### What type of PR is it?
Improvement

### Todos
* [x] - Implement cancel method

### What is the Jira issue?
https://issues.apache.org/jira/browse/ZEPPELIN-1632

### How should this be tested?
Execute a long running flink streaming job.
The "senv.execute()" method is blocking, therefore the paragraph remains in the state running. Then click cancel and the job will be canceled.

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: Philipp <philippgrulich@hotmail.de>

Closes #1607 from PhilippGrulich/flink_cancel_job and squashes the following commits:

9d00190 [Philipp] add the possibility to cancel flink jobs in local mode