[ZEPPELIN-1689] enable Power architecture
### What is this PR for?
A few sentences describing the overall goals of the pull request's commits.

this patch is platform independent and should not effect Intel architecture. We are simply upgrading the version of fronend-maven-plugin, nodejs, phantomjs-launcher to versions that are supported by Power architecture as well.

### What type of PR is it?
[ Feature ]

### Todos
* [ ] - Task

### What is the Jira issue?

https://issues.apache.org/jira/browse/ZEPPELIN-1689

### How should this be tested?
should test on x86 for any regression, highly unlikely.
OpenPOWER foundation team has already verified the patch on Power as part of Apache Bigtop project.

### Screenshots (if appropriate)

### Questions:
* Does the licenses files need update?
* Is there breaking changes for older versions?
* Does this needs documentation?

Author: amir sanjar <afsanjar@gmail.com>

Closes #1666 from asanjar/master and squashes the following commits:

2f92845 [amir sanjar] ZEPPELIN-1689: enable Power architecture