[ZEPPELIN-1729] With CRLF line endings, Livy gets HTTP 500 errors
### What is this PR for?
When you paste text on Windows, you may inadvertedly add some Carriage Return
control characters along with it. When this happens in a Livy paragraph
in Zeppelin 0.6.0, execution of that paragraph will fail with a "500 Internal
Server Error" message. This is because `LivyHelper` does not properly sanitise
its input when generating a JSON payload for the Livy server. This PR takes
the subset for the resolution of [ZEPPELIN-1430](https://issues.apache.org/jira/browse/ZEPPELIN-1430/)
that should resolve this issue.

### What type of PR is it?
Bug Fix

### What is the Jira issue?
<https://issues.apache.org/jira/browse/ZEPPELIN-1729/>

### How should this be tested?

Paste the following code in a new paragraph, ensuring that lines are ended by
CRLF rather than LF:

    %livy.pyspark
    import datetime
    import json

This can be easily realised on Windows, by copy-pasting this code from Notepad.
Then, try to execute the paragraph. Without the fix, a "500 Internal Server Error"
message should appear. With the fix, the paragraph will successfully execute
(and do nothing interesting).

### Questions:
* Does the licenses files need update? No.
* Is there breaking changes for older versions? No.
* Does this needs documentation? No.