[ZEPPELIN-1753] Fix blank notename issue
### What is this PR for?
Note ID is displayed on action bar after pressing enter key with blank notename.
This is because the scope variable for notename is used directly as the model of input field and it doesn't revert the blank notename back to its existing name even though the notename was actually not updated to the blank value. (Temporarily lost notename from scope)
To resolve this issue, I added a variable to `ElasticInputCtrl` for the model of input field.
This change can always make the notename valid.

### What type of PR is it?
Bug Fix

### Todos

### What is the Jira issue?
https://issues.apache.org/jira/browse/ZEPPELIN-1753

### How should this be tested?
Input blank notename on action bar and..
- Check the notename on action bar
- Click 'Export this note' button and check the downloaded json filename.

### Screenshots (if appropriate)
![blank-notename](https://cloud.githubusercontent.com/assets/17305893/20969808/07b6a970-bcce-11e6-9457-e264a2bb0f92.gif)

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: Sangwoo Lee <marchpig87@gmail.com>

Closes #1733 from marchpig/ZEPPELIN-1753 and squashes the following commits:

25c2ebf [Sangwoo Lee] Fix blank notename issue