[HOTFIX][ZEPPELIN-1779]Flaky test: PySparkInterpreterMatplotlibTest
### What is this PR for?
The matplotlib tests have stalled as a result of conda downloading version 2.0.0b1. While this was an old beta version and the issue has been fixed upstream (https://github.com/conda-forge/matplotlib-feedstock/issues/86), we still have not tested the inline backend for 2.0.0 which will be a release with a lot of major changes, hence this PR should be made anyway. I am also partially undoing one of the changes made in #1709 which suppressed the output of the dependency installation from conda. This change was very puzzling to me since dependency info isn't that verbose. I ended up determining the cause of this issue because I reverted this change and was then able to see the dependency versions in the log. I could see suppressing this being necessary for the R packages in case they aren't cached and a wall of compiler output gets printed out, but I see no real justification to mute it for python packages, especially given that they will only be installed for two short profiles.

### What type of PR is it?
Hot Fix

### What is the Jira issue?
[ZEPPELIN-1779](https://issues.apache.org/jira/browse/ZEPPELIN-1779)

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: Alex Goodman <agoodm@users.noreply.github.com>

Closes #1741 from agoodm/ZEPPELIN-1779 and squashes the following commits:

1211a2f [Alex Goodman] Ensure matplotlib version is 1.5.3