[ZEPPELIN-1763] Prevent duplicate $scope.$on('setNoteMenu') calls
### What is this PR for?
It's to fix the issue 'ZEPPELIN-1763' (Zeppelin hangs if I repeatedly select and deselect note name).
When the note name was changed, there was a problem that the 'setNoteName' listener was registered in duplicate and occupied resources.
To fix it, I moved the function that registers the listener out of initNoteBook()

### What type of PR is it?
[Bug Fix]

### Todos

### What is the Jira issue?
[ZEPPELIN-1763](https://issues.apache.org/jira/browse/ZEPPELIN-1763?filter=-2)

### How should this be tested?
Repeat changing note name.

### Screenshots (if appropriate)

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

The listners are added in duplicate on $scope.$on('setNoteMenu') calls.

Author: Myoungdo Park <cuspymd@gmail.com>

Closes #1770 from cuspymd/note-name-pr2 and squashes the following commits:

2616277 [Myoungdo Park] Fix a style error
99a1031 [Myoungdo Park] Prevent duplicate $scope.$on('setNoteMenu') calls