[ZEPPELIN-1239] Google WebFont: goog-webfont-dl Grunt task warning and build error
### What is this PR for?
It is to fix a zeppelin-web build error in Windows 10.
To fix it, I changed google webfont download module of grunt.

### What type of PR is it?
[Bug Fix]

### Todos

### What is the Jira issue?
[ZEPPELIN-1239](https://issues.apache.org/jira/browse/ZEPPELIN-1239)

### How should this be tested?
zeppelin-web build in Windows 10

### Screenshots (if appropriate)

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

'spawn UNKNOWN' error occurs on zeppelin-web build processing in windows 10.
It due to the function 'grunt.util.spawn()' which is called by the module 'grunt-goog-webfont-dl'.
So I changed this module to 'grunt-google-fonts'.

Author: Myoungdo Park <cuspymd@gmail.com>

Closes #1731 from cuspymd/google-fonts and squashes the following commits:

e1093fc [Myoungdo Park] Revert dangle-commas deletion
7fb336b [Myoungdo Park] Move 'grunt googlefonts' into 'postinstall'
1fea68f [Myoungdo Park] Change google font downloader to fix a build error in win 7