ZEPPELIN-1918. Fix build with Spark 2.1.0
### What is this PR for?
It's my misunderstanding of `SPARK_VER` in travis. It is only used for downloading spark distribution. We need to specify spark profile explicitly for building with specific version of spark. This PR add new profile for spark 2.1 and fix the build issue with spark 2.1.0 because `SecurityManager` changes its constructor signature in spark 2.1

### What type of PR is it?
[Bug Fix ]

### Todos
* [ ] - Task

### What is the Jira issue?
* https://issues.apache.org/jira/browse/ZEPPELIN-1918

### How should this be tested?
Build with spark 2.1.0 and tested it manually as below screenshot.

### Screenshots (if appropriate)
![image](https://cloud.githubusercontent.com/assets/164491/21797414/d586aa04-d749-11e6-8c6f-3b12e9e2ae2d.png)

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: Jeff Zhang <zjffdu@apache.org>

Closes #1880 from zjffdu/ZEPPELIN-1918 and squashes the following commits:

8f87459 [Jeff Zhang] update release script
34772e0 [Jeff Zhang] update build doc for spark 2.1
5efdb11 [Jeff Zhang] fix unit test and address comments
def502f [Jeff Zhang] ZEPPELIN-1918. Fix build with Spark 2.1.0