[HOTFIX][ZEPPELIN-1980] - Test and update CI for matplotlib 2.0.0
### What is this PR for?
Matplotlib 2.0.0 was just released. It has introduced some major changes including some to the `rcParams` which makes it incompatible with Zeppelin's built-in plotting backend. This PR updates the backend config for 2.0.0 as well as the CI tests.

### What type of PR is it?
Hot Fix

### What is the Jira issue?
[ZEPPELIN-1980](https://issues.apache.org/jira/browse/ZEPPELIN-1980)

### How should this be tested?
Run the matplotlib tutorial notebook with matplotlib 2.0.0 installed.

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

If possible, this should be included as a hotfix for 0.7.0.

Author: Alex Goodman <agoodm@users.noreply.github.com>

Closes #1912 from agoodm/ZEPPELIN-1980 and squashes the following commits:

202548c [Alex Goodman] Use figure.dpi instead of savefig.dpi
2678359 [Alex Goodman] Make CI install latest matplotlib

(cherry picked from commit 7941d37518a9393435680833c0828b13c77dd863)
Signed-off-by: Mina Lee <minalee@apache.org>