[ZEPPELIN-1465] Add an option to allow S3 server-side encryption
### What is this PR for?
Provide a configuration option that will cause the S3 Notebook repo to request server-side encryption of saved notebooks.

### What type of PR is it?
Improvement

### What is the Jira issue?
https://issues.apache.org/jira/browse/ZEPPELIN-1465

### How should this be tested?
Enable the configuration option, save a notebook in zeppelin, and confirm in the AWS S3 Console that the related file was saved with AES-256 encryption on the server-side.  (Properties tab, Detail section)

### Questions:
* Does the licenses files need update?
No

* Is there breaking changes for older versions?
No.

* Does this needs documentation?
I added mentions of the new option in existing documentation.

Thank you!

Author: Jeff Plourde <jplourde@cyft.io>

Closes #1969 from jeff-cyft/s3_sse and squashes the following commits:

26f5264 [Jeff Plourde] code style - remove tab
3c657ac [Jeff Plourde] Configuration option to request S3 SSE when notebooks are saved.