When gear icon is clicked, scroll the page to the top to show the interpreter bindings
### What is this PR for?
When someone is browsing a notebook and they click the *gear* button of interpreter bindings, the bindings do come up, but the person might not realize since it opens on the top of the page, and they don't get any feedback that the bindings have opened. We should scroll up the page to show the bindings. This will also act as the feedback of clicking the *gear* button.

### What type of PR is it?
[Improvement]

### Todos
* None

### What is the Jira issue?
https://issues.apache.org/jira/browse/ZEPPELIN-2047

### How should this be tested?
Create few (say 10) paragraphs and scroll down to the last paragraph. Now click the *gear* icon meant to open the interpreter bindings. The view should have scrolled up to show the bindings.

### Screenshots (if appropriate)

![zeppelin-2047](https://cloud.githubusercontent.com/assets/4542030/22618870/e724a1f8-eb0d-11e6-9395-949791469659.gif)

### Questions:
* Does the licenses files need update?
No
* Is there breaking changes for older versions?
No
* Does this needs documentation?
No. Its a usability improvement.

Author: Beria <beria@qubole.com>

Closes #1971 from beriaanirudh/master and squashes the following commits:

e7b8e51 [Beria] scroll gracefully
a0f521a [Beria] When gear icon is clicked, scroll the page to the top to show the intepreter bindings