[ZEPPELIN-2116]. livy interpreter needs to restart after previous session timeout in secured cluster
What is this PR for?

It throw different exception in secured cluster. This PR fix the issue in secured cluster

What type of PR is it?

[Bug Fix]

Todos

 - Task
What is the Jira issue?

https://issues.apache.org/jira/browse/ZEPPELIN-2116
How should this be tested?

Tested manually.

Questions:

Does the licenses files need update? No
Is there breaking changes for older versions? No
Does this needs documentation? no

Author: Jeff Zhang <zjffdu@apache.org>

Closes #2021 from zjffdu/ZEPPELIN-2116 and squashes the following commits:

05e6a3c [Jeff Zhang] address comments
96a7cfe [Jeff Zhang] update
63c693b [Jeff Zhang] add more acurrate detection for session expire
7949b51 [Jeff Zhang] [ZEPPELIN-2116]. livy interpreter needs to restart after previous session timeout in secured cluster