[ZEPPELIN-2167] User with insufficient privileges can still restore files by renaming files in/out of Trash
### What is this PR for?
User with insufficient privileges can still restore files by renaming files in/out of Trash

### What type of PR is it?
[Bug Fix]

### What is the Jira issue?
* [ZEPPELIN-2167](https://issues.apache.org/jira/browse/ZEPPELIN-2167)

### How should this be tested?
Steps to reproduce:
 - Create a notebook "test_nb" as bob.
 - Delete the notebook
 - Login as mary and try restoring "test_nb" from Trash folder. The system correctly complains of insufficient privileges.
 - Open the "test_nb" notebook from Trash folder. The notebook opens with title "~Trash/test_nb".
 - Edit the title and remove the prefix "~Trash".

If you now look at the list of notebooks there is no file "test_nb" in Trash.
Interestingly when you try and delete the recently moved file from Trash it complains that mary does not have privileges to delete it. Edit the title of that notebook to "~Trash/test_nb" and it goes back to Trash folder.

### Questions:
* Does the licenses files need update? N/A
* Is there breaking changes for older versions? N/A
* Does this needs documentation? N/A

Author: Prabhjyot Singh <prabhjyotsingh@gmail.com>

Closes #2086 from prabhjyotsingh/ZEPPELIN-2167 and squashes the following commits:

07221e0 [Prabhjyot Singh] updateNoteName should use renameNote function